Split page cache tracer into page cache tracer and page cursor tracer. Cleanup interfaces. Tests are in best case compilable.
