+++ community/io/src/test/java/org/neo4j/io/pagecache/tracing/DefaultPageCacheTracerTest.java
@@ -19,12 +19,109 @@
  */
 package org.neo4j.io.pagecache.tracing;
 
-public class DefaultPageCacheTracerTest extends PageCacheTracerTest
+import org.junit.Before;
+import org.junit.Test;
+
+import java.io.File;
+import java.io.IOException;
+
+import org.neo4j.io.pagecache.PageSwapper;
+
+import static org.hamcrest.Matchers.is;
+import static org.junit.Assert.assertThat;
+
+public class DefaultPageCacheTracerTest
 {
-    @Override
-    protected DefaultPageCacheTracer createTracer()
+    private PageCacheTracer tracer;
+    private PageSwapper swapper;
+
+    @Before
+    public void setUp()
     {
-        DefaultPageCacheTracer.enablePinUnpinTracing();
-        return new DefaultPageCacheTracer();
+        tracer = new DefaultPageCacheTracer();
+        swapper = new DummyPageSwapper( "filename" );
+    }
+
+    private void assertCounts( long pins, long unpins, long faults, long evictions, long evictionExceptions,
+                               long flushes, long bytesRead, long bytesWritten, long filesMapped, long filesUnmapped )
+    {
+        assertThat( "pins", tracer.pins(), is( pins ) );
+        assertThat( "unpins", tracer.unpins(), is( unpins ) );
+        assertThat( "faults", tracer.faults(), is( faults ) );
+        assertThat( "evictions", tracer.evictions(), is( evictions ) );
+        assertThat( "evictionExceptions", tracer.evictionExceptions(), is( evictionExceptions ) );
+        assertThat( "flushes", tracer.flushes(), is( flushes ) );
+        assertThat( "bytesRead", tracer.bytesRead(), is( bytesRead ) );
+        assertThat( "bytesWritten", tracer.bytesWritten(), is( bytesWritten ) );
+        assertThat( "filesMapped", tracer.filesMapped(), is( filesMapped ) );
+        assertThat( "filesUnmapped", tracer.filesUnmapped(), is( filesUnmapped ) );
+    }
+
+    @Test
+    public void mustCountEvictions()
+    {
+        try ( EvictionRunEvent evictionRunEvent = tracer.beginPageEvictions( 2 ) )
+        {
+            try ( EvictionEvent evictionEvent = evictionRunEvent.beginEviction() )
+            {
+                FlushEvent flushEvent = evictionEvent.flushEventOpportunity().beginFlush( 0, 0, swapper );
+                flushEvent.addBytesWritten( 12 );
+                flushEvent.done();
+            }
+
+            try ( EvictionEvent evictionEvent = evictionRunEvent.beginEviction() )
+            {
+                FlushEvent flushEvent = evictionEvent.flushEventOpportunity().beginFlush( 0, 0, swapper );
+                flushEvent.addBytesWritten( 12 );
+                flushEvent.done();
+                evictionEvent.threwException( new IOException() );
+            }
+
+            try ( EvictionEvent evictionEvent = evictionRunEvent.beginEviction() )
+            {
+                FlushEvent flushEvent = evictionEvent.flushEventOpportunity().beginFlush( 0, 0, swapper );
+                flushEvent.addBytesWritten( 12 );
+                flushEvent.done();
+                evictionEvent.threwException( new IOException() );
+            }
+
+            evictionRunEvent.beginEviction().close();
+        }
+
+        assertCounts( 0, 0, 0, 4, 2, 3, 0, 36, 0, 0 );
+    }
+
+    @Test
+    public void mustCountFileMappingAndUnmapping()
+    {
+        tracer.mappedFile( new File( "a" ) );
+
+        assertCounts( 0, 0, 0, 0, 0, 0, 0, 0, 1, 0 );
+
+        tracer.unmappedFile( new File( "a" ) );
+
+        assertCounts( 0, 0, 0, 0, 0, 0, 0, 0, 1, 1 );
+    }
+
+    @Test
+    public void mustCountFlushes()
+    {
+        try ( MajorFlushEvent cacheFlush = tracer.beginCacheFlush() )
+        {
+            cacheFlush.flushEventOpportunity().beginFlush( 0, 0, swapper ).done();
+            cacheFlush.flushEventOpportunity().beginFlush( 0, 0, swapper ).done();
+            cacheFlush.flushEventOpportunity().beginFlush( 0, 0, swapper ).done();
+        }
+
+        assertCounts( 0, 0, 0, 0, 0, 3, 0, 0, 0, 0 );
+
+        try ( MajorFlushEvent fileFlush = tracer.beginFileFlush( swapper ) )
+        {
+            fileFlush.flushEventOpportunity().beginFlush( 0, 0, swapper ).done();
+            fileFlush.flushEventOpportunity().beginFlush( 0, 0, swapper ).done();
+            fileFlush.flushEventOpportunity().beginFlush( 0, 0, swapper ).done();
+        }
+
+        assertCounts( 0, 0, 0, 0, 0, 6, 0, 0, 0, 0 );
     }
 }
