+++ community/io/src/test/java/org/neo4j/test/LinearHistoryPageCursorTracer.java
@@ -17,35 +17,20 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package org.neo4j.io.pagecache.tracing;
+package org.neo4j.test;
 
-/**
- * An eviction run is started when the page cache has determined that it
- * needs to evict a batch of pages. The dedicated eviction thread is
- * mostly sleeping when it is not performing an eviction run.
- */
-public interface EvictionRunEvent extends AutoCloseablePageCacheTracerEvent
+import org.neo4j.io.pagecache.PageSwapper;
+import org.neo4j.io.pagecache.tracing.PinEvent;
+import org.neo4j.io.pagecache.tracing.cursor.PageCursorTracer;
+
+public class LinearHistoryPageCursorTracer implements PageCursorTracer
 {
-    /**
-     * An EvictionRunEvent that does nothing other than return the EvictionEvent.NULL.
-     */
-    EvictionRunEvent NULL = new EvictionRunEvent()
+
+    //TODO:
+    @Override
+    public PinEvent beginPin( boolean writeLock, long filePageId, PageSwapper swapper )
     {
+        return PinEvent.NULL;
+    }
 
-        @Override
-        public EvictionEvent beginEviction()
-        {
-            return EvictionEvent.NULL;
-        }
-
-        @Override
-        public void close()
-        {
-        }
-    };
-
-    /**
-     * An eviction is started as part of this eviction run.
-     */
-    EvictionEvent beginEviction();
 }
